Improving the class-to-table mapping

+++ Chapter 12/Northwind.EntityModels.Sqlite/OrderDetail.cs
@@ -14,27 +14,27 @@ namespace Northwind.EntityModels.Sqlite;
 [Index("ProductId", Name = "ProductsOrder_Details")]
 public partial class OrderDetail
 {
-    [Key]
-    [Column(TypeName = "INT")]
-    public int OrderId { get; set; }
+	[Key]
+	[Column(TypeName = "INT")]
+	public int OrderId { get; set; }
 
-    [Key]
-    [Column(TypeName = "INT")]
-    public int ProductId { get; set; }
+	[Key]
+	[Column(TypeName = "INT")]
+	public int ProductId { get; set; }
 
-    [Column(TypeName = "money")]
-    public double UnitPrice { get; set; }
+	[Column(TypeName = "money")]
+	public decimal UnitPrice { get; set; }
 
-    [Column(TypeName = "smallint")]
-    public short Quantity { get; set; }
+	[Column(TypeName = "smallint")]
+	public short Quantity { get; set; }
 
-    public double Discount { get; set; }
+	public double Discount { get; set; }
 
-    [ForeignKey("OrderId")]
-    [InverseProperty("OrderDetails")]
-    public virtual Order Order { get; set; } = null!;
+	[ForeignKey("OrderId")]
+	[InverseProperty("OrderDetails")]
+	public virtual Order Order { get; set; } = null!;
 
-    [ForeignKey("ProductId")]
-    [InverseProperty("OrderDetails")]
-    public virtual Product Product { get; set; } = null!;
+	[ForeignKey("ProductId")]
+	[InverseProperty("OrderDetails")]
+	public virtual Product Product { get; set; } = null!;
 }
